fix: improve eslint plugin perf
Do not send the AST to the worker thread.

This changed the local test from 8.4 seconds to 6 seconds.
