+++ packages/cspell-eslint-plugin/src/plugin/tsconfig.json
@@ -10,7 +10,7 @@
     "include": ["."],
     "references": [
         { "path": "../common/tsconfig.json" },
-        { "path": "../worker/tsconfig.json" },
+        { "path": "../spellCheckAST/tsconfig.json" },
         { "path": "../generated/tsconfig.json" }
     ]
 }
