+++ packages/cspell-eslint-plugin/src/spellCheckAST/ASTNode.ts
