+++ packages/cspell-eslint-plugin/src/spellCheckAST/customScopes.cts
