+++ packages/cspell-eslint-plugin/src/synckit/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+        "module": "CommonJS",
+        "moduleResolution": "Node",
+        "tsBuildInfoFile": "../../dist/synckit/compile.tsbuildInfo",
+        "rootDir": ".",
+        "outDir": "../../dist/synckit"
+    },
+    "include": ["."],
+    "references": []
+}
