+++ packages/cspell-eslint-plugin/src/test/tsconfig.json
@@ -7,5 +7,5 @@
         "outDir": "../../dist/test"
     },
     "include": ["."],
-    "references": [{ "path": "../plugin/tsconfig.json" }, { "path": "../worker/tsconfig.json" }]
+    "references": [{ "path": "../plugin/tsconfig.json" }]
 }
