--- packages/cspell-eslint-plugin/src/worker/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../../tsconfig.base.json",
-    "compilerOptions": {
-        "tsBuildInfoFile": "../../dist/worker/compile.tsbuildInfo",
-        "allowJs": true,
-        "rootDir": ".",
-        "outDir": "../../dist/worker"
-    },
-    "include": ["."],
-    "references": [{ "path": "../common/tsconfig.json" }, { "path": "../test-util/tsconfig.json" }]
-}
