+++ packages/cspell-eslint-plugin/tsconfig.json
@@ -7,7 +7,8 @@
     "references": [
         { "path": "./src/test/tsconfig.json" },
         { "path": "./src/generated/tsconfig.json" },
-        { "path": "./src/worker/tsconfig.json" },
+        { "path": "./src/spellCheckAST/tsconfig.json" },
+        { "path": "./src/synckit/tsconfig.json" },
         { "path": "./src/common/tsconfig.json" },
         { "path": "./src/plugin/tsconfig.json" }
     ]
